% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{as_demogdata}
\alias{as_demogdata}
\title{Convert the death rate humble data to demogdata}
\usage{
as_demogdata(data, country = NULL)
}
\arguments{
\item{data}{A humble object from \code{hmd_data()} with
\code{stats = c("death_rate", "exposures_to_risk")}.}

\item{country}{A single country within the data. If \code{data} contains only
a single country, then there is no need to supply this.}
}
\description{
Note that data should containly only one country or supply the country to
select in the \code{country} argument.
}
