% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Utility.R
\name{generate_default_qx}
\alias{generate_default_qx}
\title{Generate Default Death Probabilities}
\usage{
generate_default_qx(init_age, sex = "F", closure_age = 130)
}
\arguments{
\item{init_age}{the initial age of the path}

\item{sex}{character string denoting the gender of individuals, "F" for female and "M" for male}

\item{closure_age}{maximum life span}
}
\value{
a vector of 1-year death probabilities corresponding to the specified
initial and closure age
}
\description{
Generates default death probabilities to be used for simulating paths if
necessary.
}
\keyword{internal}
