% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{hmd_data}
\alias{hmd_data}
\title{Fetch the data from the Human Mortality Database}
\usage{
hmd_data(
  country,
  stats = "death_rate",
  year_range = 1,
  age_range = 1,
  user = NULL
)
}
\arguments{
\item{country}{The country code to fetch the data for. The input may be
a named or unnamed character vector. If the vector is named, then the
name is used as the label.}

\item{stats}{The statistics to extract. The available statistics are:
"birth", "death", "life_expectancy", "exposure_to_risk", "population",
and "life_tables".}

\item{user}{Either a named list of \code{username} and \code{password} to access the
HMD or if \code{NULL} then the username and password set by \code{hmd_user} is
used instead. The recommended approach is just to set the user session
with \code{hmd_user()}.}

\item{range_year, range_age}{A single integer indicating the period or
cohort year/age range.}
}
\value{
A humble object.
}
\description{
Fetch the data from the Human Mortality Database
}
\seealso{
hmd_birth
}
