% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{hmd_session}
\alias{hmd_session}
\title{Set the username and password to access the Human Mortality Database}
\usage{
hmd_session(
  username = Sys.getenv("HMD_USERNAME"),
  password = Sys.getenv("HMD_PASSWORD")
)
}
\arguments{
\item{username}{The username (typically your email) that you signed up for HMD.}

\item{password}{The password used to access HMD.}
}
\description{
Set the username and password to access the Human Mortality Database
}
