% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz-makeham.R
\name{life_expectancy_GM}
\alias{life_expectancy_GM}
\title{Calculate Gompertz-Makeham Life Expectancy}
\usage{
life_expectancy_GM(t, A, R, alpha)
}
\arguments{
\item{t}{\code{numeric scalar or vector} corresponding to the age(s) \mjseqn{t}}

\item{A}{\code{numeric scalar} corresponding to the \mjseqn{A} parameter}

\item{R}{\code{numeric scalar} corresponding to the \mjseqn{R} parameter}

\item{alpha}{\code{numeric scalar} corresponding to the \mjseqn{\alpha} parameter}
}
\value{
Returns a \code{numeric scalar} giving the Gompertz-Makeham life
expectancy \mjseqn{e_{GM}(t)}
}
\description{
\loadmathjax{}
Calculates
\mjsdeqn{ \mathcal{M}_{GM}(t) =  \frac{1}{b} \exp\left( \frac{a e^{bt}}{b}
 \right ) \left( \frac{a e^{bx}}{b} \right)^{c/b}
 \Gamma\left( - \frac{c}{b}, \frac{a e^{bt}}{b} \right)
 }
with the upper incomplete gamma function
\mjsdeqn{\Gamma(\cdot, \cdot) = \int_x^\infty t^{u-1} e^{-t} dt, \; x>0,
 \quad u \in \mathbb{R}
}
which is implemented in \code{\link{upper_inc_gamma}}.
}
\examples{

# You can use a single age:
life_expectancy_GM(t = 0,
                   A = 0.00033,
                   R = 0.00014,
                   alpha = 0.11521)

# Or a vector of ages:
life_expectancy_GM(t = 0:100,
                   A = 0.00033,
                   R = 0.00014,
                   alpha = 0.11521)

}
\seealso{
\code{\link{upper_inc_gamma}}
}
