% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mortality_model.R
\docType{data}
\name{mortality_model}
\alias{mortality_model}
\title{Mortality model}
\format{a \code{\link{glm}} object}
\source{
frailtypack
}
\usage{
mortality_model
}
\description{
Simple example model that predicts mortality based on several variables
}

