% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_Rate2Rate.R
\name{rate2rate}
\alias{rate2rate}
\title{Convert mortality rates}
\usage{
rate2rate(rates, from, to)
}
\arguments{
\item{rates}{vector, matrix or 3D array of mortality rates}

\item{from}{character string representing the type of mortality rate to be converted
from. Takes the following values: "central" for central death rates, "prob"
for 1-year death probabilities, "force" for force of mortality.}

\item{to}{character string representing the type of mortality rate to be converted to.
Takes the same values as the parameter \code{from}}
}
\value{
vector, matrix or 3D array of converted mortality rates
}
\description{
Converts between central death rates, 1-year death probabilities and force of
mortality.
}
\details{
Implementation assumes a constant force of mortality for fractional ages,
resulting in the following conversions:
\deqn{q_{x, y} = 1 - e^{-\mu_{x, y}}}
\deqn{\mu_{x, y} = m_{x, y}}
\deqn{m_{x, y}  = -log(1 - q_{x, y})}
}
