% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{shiny_app}
\alias{shiny_app}
\title{Run a Shiny app displaying the survival model functions}
\usage{
shiny_app(plot_resolution = 100)
}
\arguments{
\item{plot_resolution}{\code{Numeric scalar} giving the resolution for plots}
}
\value{
Runs a \code{shiny} app
}
\description{
Run a Shiny app displaying the survival model functions
}
