% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stop_if_not_scalar}
\alias{stop_if_not_scalar}
\title{Test if an object is a scalar}
\usage{
stop_if_not_scalar(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Displays an error message informing that x should be a scalar.
}
\description{
Test if an object is a scalar
}
\keyword{1:10}
\keyword{=}
\keyword{Internal}
\keyword{stop_if_not_scalar(x)}
\keyword{x}
