% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stop_if_package_is_missing}
\alias{stop_if_package_is_missing}
\title{Test if an package is installed and fail else}
\usage{
stop_if_package_is_missing(package)
}
\arguments{
\item{package}{\code{string scalar} Giving the package to test for}
}
\value{
Displays an error message indicating that the package should be installed
}
\description{
Test if an package is installed and fail else
}
\keyword{Internal}
