% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_PriceOfRisk.R
\name{survival2rate}
\alias{survival2rate}
\title{Convert survival function to death probabilities}
\usage{
survival2rate(surv, ages, to = "prob", years = NULL)
}
\arguments{
\item{surv}{vector, matrix or 3D array of the survival function with survival time starting from 0
(on the rows) and calendar year or cohort (on the columns) and simulation number (3rd dimension)}

\item{ages}{vector of desired ages for the resulting 1-year death probabilities}

\item{to}{character string representing the type of mortality rate to be converted
to. Takes the following values: "central" for central death rates, "prob"
for 1-year death probabilities, "force" for force of mortality.}

\item{years}{optional vector of years for \code{surv}. If not supplied, then the column names
of \code{surv} will be preserved}
}
\value{
associated mortality rates in the same format as \code{surv}
}
\description{
Converts the survival function to the associated mortality rates.
}
