% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{upper_inc_gamma}
\alias{upper_inc_gamma}
\title{Upper incomplete gamma distribution}
\usage{
upper_inc_gamma(u, x)
}
\arguments{
\item{u}{\code{Numeric scalar} corresponding to \mjseqn{u}}

\item{x}{\code{Numeric scalar} corresponding to \mjseqn{x}}
}
\value{
Returns a \code{numeric scalar}
}
\description{
\loadmathjax{}
Calculates
\mjsdeqn{\int_x^\infty t^{u-1} \exp(-t) dt}

using numerical integration via \code{\link{integrate}}.
}
\examples{

upper_inc_gamma(-0.005, 2)

}
