% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertz-makeham.R
\name{force_of_mortality_GM}
\alias{force_of_mortality_GM}
\title{Calculate Gompertz-Makeham Force of Motion}
\usage{
force_of_mortality_GM(A, R, alpha, t)
}
\arguments{
\item{A}{\code{numeric scalar} corresponding to the \mjseqn{A} parameter}

\item{R}{\code{numeric scalar} corresponding to the \mjseqn{R} parameter}

\item{alpha}{\code{numeric scalar} corresponding to the \mjseqn{\alpha} parameter}

\item{t}{\code{numeric scalar or vector} corresponding to the age(s) \mjseqn{t}}
}
\value{
Returns a \code{numeric scalar} giving the Gompertz-Makeham force
of mortality \mjseqn{\mu(t)}
}
\description{
\loadmathjax{}
Calculates \mjseqn{\mu(t) = A + R e^{\alpha t}}
}
\examples{

# You can use a single age:
force_of_mortality_GM(A = 0.005,
                      R = 0.00005,
                      alpha = 0.0093,
                      t = 10)

# Or a vector of ages:
force_of_mortality_GM(A = 0.005,
                      R = 0.00005,
                      alpha = 0.0093,
                      t = 0:100)

}
