% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{hmd_session}
\alias{hmd_session}
\title{Set the username and password to access the Human Mortality Database}
\usage{
hmd_session(
  username = Sys.getenv("HMD_USERNAME"),
  password = Sys.getenv("HMD_PASSWORD")
)
}
\arguments{
\item{username}{The username (typically your email) that you signed up for HMD.}

\item{password}{The password used to access HMD.}
}
\description{
You first need to register as a user
(if you haven't already) at the Human Mortality Database.
To now set a session pass in your username and password from your account or
alternatively, you can store your username and password in the .Renviron file
containing

HMD_USERNAME=<YOUR USERNAME>

HMD_PASSWORD=<YOUR PASSWORD>
}
