% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stop_if_not_scalar}
\alias{stop_if_not_scalar}
\title{Test if an object is a scalar}
\usage{
stop_if_not_scalar(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Returns a {Logical scalar} with value \code{TRUE} if the object is a
 scalar
}
\description{
Test if an object is a scalar
}
